requirements-rsenv.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main.conda
_openmp_mutex=5.1=1_gnu.conda
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
boost=1.67.0=py37_4
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.4.26=h06a4308_0.conda
cairo=1.14.12=h8948797_3
certifi=2022.5.18.1=py37h06a4308_0.conda
cloudpickle=1.6.0=py_0
cudatoolkit=10.1.243=h6bb024c_0.conda
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2.30.0=py_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py37h295c915_0.conda
decorator=5.1.1=pyhd3eb1b0_0.conda
entrypoints=0.4=py37h06a4308_0.conda
expat=2.2.10=he6710b0_2
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.4=h5ab3b9f_0
gflags=2.2.2=he6710b0_0
glib=2.56.2=hd408876_0
glog=0.3.5=hf484d3e_1.conda
graphite2=1.3.14=h23475e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py37h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
imageio=2.9.0=py_0
intel-openmp=2020.2=254
ipykernel=6.9.1=py37h06a4308_0.conda
ipython=7.31.1=py37h06a4308_0.conda
ipython_genutils=0.2.0=pyhd3eb1b0_1.conda
jasper=2.0.14=h07fcdf6_1
jedi=0.18.1=py37h06a4308_1.conda
jpeg=9b=habf39ab_1
jupyter_client=7.2.2=py37h06a4308_0.conda
jupyter_core=4.10.0=py37h06a4308_0.conda
kiwisolver=1.2.0=py37hfd86e86_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.38=h1181459_1.conda
leveldb=1.20=hf484d3e_1.conda
libboost=1.67.0=h46d08c1_4
libedit=3.1.20210910=h7f8727e_0.conda
libffi=3.3=he6710b0_2.conda
libgcc-ng=11.2.0=h1234567_1.conda
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libgomp=11.2.0=h1234567_1.conda
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=hd408876_0
libsodium=1.0.18=h7b6447c_0.conda
libstdcxx-ng=11.2.0=h1234567_1.conda
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lmdb=0.9.24=he6710b0_0
lz4-c=1.9.2=heb0550a_3
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py37h817c723_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2.conda
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.0=py37hd6b4f25_0
ncurses=6.3=h7f8727e_2.conda
nest-asyncio=1.5.5=py37h06a4308_0.conda
networkx=2.5=py_0
ninja=1.10.2=hff7bd54_1.conda
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
olefile=0.46=py37_0
openssl=1.1.1o=h7f8727e_0.conda
pandas=1.1.3=py37he6710b0_0
parso=0.8.3=pyhd3eb1b0_0.conda
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3.conda
pickleshare=0.7.5=pyhd3eb1b0_1003.conda
pillow=8.0.0=py37h9a89aac_0
pip=21.2.2=py37h06a4308_0.conda
pixman=0.40.0=h7b6447c_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0.conda
protobuf=3.13.0.1=py37he6710b0_1
ptyprocess=0.7.0=pyhd3eb1b0_2.conda
py-boost=1.67.0=py37h04863e7_4
py-opencv=3.4.2=py37hb342d67_1
pygments=2.11.2=pyhd3eb1b0_0.conda
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h22d08a2_1
python=3.7.13=h12debd9_0.conda
python-dateutil=2.8.2=pyhd3eb1b0_0.conda
python-gflags=3.1.2=py37_0
python-leveldb=0.201=py37he6710b0_0
pytorch=1.7.1=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2020.1=py_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.3.1=py37h7b6447c_1
pyzmq=22.3.0=py37h295c915_2.conda
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1.conda
scikit-image=0.17.2=py37hdf5156a_0
scipy=1.5.2=py37h0b6359f_0
setuptools=61.2.0=py37h06a4308_0.conda
sip=4.19.24=py37he6710b0_0
six=1.16.0=pyhd3eb1b0_1.conda
snappy=1.1.8=he6710b0_0
sqlite=3.38.3=hc218d9a_0.conda
tifffile=2020.10.1=py37hdd07704_2
tk=8.6.12=h1ccaba5_0.conda
toolz=0.11.1=py_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu101
tornado=6.1=py37h27cfd23_0.conda
traitlets=5.1.1=pyhd3eb1b0_0.conda
typing_extensions=3.10.0.2=pyh06a4308_0.conda
tzdata=2020f=h52ac0ba_0
wcwidth=0.2.5=pyhd3eb1b0_0.conda
wheel=0.37.1=pyhd3eb1b0_0.conda
xz=5.2.5=h7f8727e_1.conda
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0.conda
zlib=1.2.12=h7f8727e_2.conda
zstd=1.4.4=h0b5b093_3
